feat: Implement dataset-specific metadata bar component
