+++ superset-frontend/src/components/Chart/DrillDetail/DrillDetailPane.tsx
@@ -17,23 +17,23 @@
  * under the License.
  */
 import React, {
-  useState,
+  ReactElement,
+  useCallback,
   useEffect,
   useMemo,
-  useCallback,
   useRef,
-  ReactElement,
+  useState,
 } from 'react';
 import { useSelector } from 'react-redux';
 import {
   BinaryQueryObjectFilterClause,
   css,
   ensureIsArray,
+  GenericDataType,
+  JsonObject,
+  QueryFormData,
   t,
   useTheme,
-  QueryFormData,
-  JsonObject,
-  GenericDataType,
 } from '@superset-ui/core';
 import { useResizeDetector } from 'react-resize-detector';
 import Loading from 'src/components/Loading';
@@ -43,16 +43,12 @@ import TimeCell from 'src/components/Table/cell-renderers/TimeCell';
 import { EmptyStateMedium } from 'src/components/EmptyState';
 import { getDatasourceSamples } from 'src/components/Chart/chartAction';
 import Table, { ColumnsType, TableSize } from 'src/components/Table';
-import MetadataBar, {
-  ContentType,
-  MetadataType,
-} from 'src/components/MetadataBar';
-import Alert from 'src/components/Alert';
-import { useApiV1Resource } from 'src/hooks/apiResources';
 import HeaderWithRadioGroup from 'src/components/Table/header-renderers/HeaderWithRadioGroup';
 import TableControls from './DrillDetailTableControls';
 import { getDrillPayload } from './utils';
-import { Dataset, ResultsPage } from './types';
+import { ResultsPage } from './types';
+import { useDatasetMetadataBar } from '../../MetadataBar/dataset/useDatasetMetadataBar';
+import { ResourceStatus } from '../../../hooks/apiResources/apiResources';
 
 const PAGE_SIZE = 50;
 
@@ -106,6 +102,8 @@ export default function DrillDetailPane({
     [formData.datasource],
   );
 
+  const { MetadataBar, status: metadataBarStatus } =
+    useDatasetMetadataBar(datasourceId);
   // Get page of results
   const resultsPage = useMemo(() => {
     const nextResultsPage = resultsPages.get(pageIndex);
@@ -261,11 +259,9 @@ export default function DrillDetailPane({
     resultsPages,
   ]);
 
-  // Get datasource metadata
-  const response = useApiV1Resource<Dataset>(`/api/v1/dataset/${datasourceId}`);
-
   const bootstrapping =
-    (!responseError && !resultsPages.size) || response.status === 'loading';
+    (!responseError && !resultsPages.size) ||
+    metadataBarStatus === ResourceStatus.LOADING;
 
   let tableContent = null;
   if (responseError) {
@@ -308,76 +304,9 @@ export default function DrillDetailPane({
     );
   }
 
-  const metadata = useMemo(() => {
-    const { status, result } = response;
-    const items: ContentType[] = [];
-    if (result) {
-      const {
-        changed_on_humanized,
-        created_on_humanized,
-        description,
-        table_name,
-        changed_by,
-        created_by,
-        owners,
-      } = result;
-      const notAvailable = t('Not available');
-      const createdBy =
-        `${created_by?.first_name ?? ''} ${
-          created_by?.last_name ?? ''
-        }`.trim() || notAvailable;
-      const modifiedBy = changed_by
-        ? `${changed_by.first_name} ${changed_by.last_name}`
-        : notAvailable;
-      const formattedOwners =
-        owners.length > 0
-          ? owners.map(owner => `${owner.first_name} ${owner.last_name}`)
-          : [notAvailable];
-      items.push({
-        type: MetadataType.TABLE,
-        title: table_name,
-      });
-      items.push({
-        type: MetadataType.LAST_MODIFIED,
-        value: changed_on_humanized,
-        modifiedBy,
-      });
-      items.push({
-        type: MetadataType.OWNER,
-        createdBy,
-        owners: formattedOwners,
-        createdOn: created_on_humanized,
-      });
-      if (description) {
-        items.push({
-          type: MetadataType.DESCRIPTION,
-          value: description,
-        });
-      }
-    }
-    return (
-      <div
-        css={css`
-          display: flex;
-          margin-bottom: ${theme.gridUnit * 4}px;
-        `}
-      >
-        {status === 'complete' && (
-          <MetadataBar items={items} tooltipPlacement="bottom" />
-        )}
-        {status === 'error' && (
-          <Alert
-            type="error"
-            message={t('There was an error loading the dataset metadata')}
-          />
-        )}
-      </div>
-    );
-  }, [response, theme.gridUnit]);
-
   return (
     <>
-      {!bootstrapping && metadata}
+      {!bootstrapping && <MetadataBar />}
       {!bootstrapping && (
         <TableControls
           filters={filters}
